BLD: No python variants for ARM

+++ recipe/build.sh
@@ -11,11 +11,14 @@
 mkdir -p "${PREFIX}/${targetsDir}"
 mv -v extras/Debugger/include "${PREFIX}/${targetsDir}"
 
-rm bin/cuda-gdb
-if [[ ${PY_VER:-0} == "0" ]]; then
-    mv -v "bin/cuda-gdb-minimal" bin/cuda-gdb
-else
-    mv -v "bin/cuda-gdb-python${PY_VER}-tui" bin/cuda-gdb
+# No python support for the debugger in non-x86 platforms on CUDA 12.9
+if [[ ${target_platform} == "linux-64" ]]; then
+    rm bin/cuda-gdb
+    if [[ ${PY_VER:-0} == "0" ]]; then
+        mv -v "bin/cuda-gdb-minimal" bin/cuda-gdb
+    else
+        mv -v "bin/cuda-gdb-python${PY_VER}-tui" bin/cuda-gdb
+    fi
 fi
 
 for i in `ls`; do
